fix: load worker release environment before runtime
